+++ tools/bazar/handlers/page/api_put.php
@@ -1,15 +1,17 @@
 <?php
 
 use YesWiki\Bazar\Service\FicheManager;
+use YesWiki\Core\Service\ApiService;
 
 if (!defined("WIKINI_VERSION")) {
     die("acc&egrave;s direct interdit");
 }
 
 $ficheManager = $this->services->get(FicheManager::class);
+$apiService = $this->services->get(ApiService::class);
 
 if ($ficheManager->isFiche($this->GetPageTag())) {
-    if ($this->api->isAuthorized()) {
+    if ($apiService->isAuthorized()) {
         $semantic = true; // strpos($_SERVER['CONTENT_TYPE'], 'application/ld+json') !== false;
 
         $_POST['id_fiche'] = $this->GetPageTag();
